feat($http): allow overriding the XSRF header and cookie name
Add 'xsrfCookie' and 'xsrfHeader' property to $httpProvider.defaults and http
config object, which give the name of the cookie the XSRF token is found in,
and the name of the header it is sent in, respectively.
This allows interop with servers with built-in XSRF support that use different
names.
The defaults match the current hard-coded values of 'XSRF-TOKEN' and
'X-XSRF-TOKEN'.